# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: TD Flask API
  version: 1.0.0
  description: API documentation for TD Flask application
servers:
  - url: http://localhost:5000
paths:
...
  /video/{id}:
    get:
      summary: Generate a presigned URL for a video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Presigned URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
        '404':
          description: Video not found
  /set_video_status/{id}:
    post:
      summary: Set the status of a video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
      responses:
        '201':
          description: Video status updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  video_id:
                    type: string
                  original_status:
                    type: string
                  new_status:
                    type: string
        '400':
          description: Invalid status provided
        '404':
          description: Video not found
  /confirm_upload/{id}:
    post:
      summary: Confirm the upload of a video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Upload success
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
        '404':
          description: Video not found
  /check_video_exists/{id}:
    get:
      summary: Check if a video exists
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Video existence status
          content:
            application/json:
              schema:
                type: object
                properties:
                  exists:
                    type: boolean
...